+++ src/ranking.hpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-bool compare(const pair<string,long> &a,const pair<string,long> &b);
+bool compare(const pair<string,float> &a,const pair<string,float> &b);
 void ranking(Tempos binaryTree,Tempos avlTree,Tempos redblackTree,Tempos Vetor,Tempos unorderedMap,Tempos Map, int grandeza, int metodo);
