+++ src/vetor_map.hpp
@@ -6,16 +6,16 @@
 #include <map>
 #include "base.hpp"
 
-long insercaoUnorderedMap(unordered_map<double,Record> *hash,vector<double> numeros, int x);
-long buscaUnorderedMap(unordered_map<double, Record> *hash,vector<double> busca);
-long remocaoUnorderedMap(unordered_map<double, Record> *hash,vector<double> busca);
+float insercaoUnorderedMap(unordered_map<double,Record> *hash,vector<double> numeros, int x);
+float buscaUnorderedMap(unordered_map<double, Record> *hash,vector<double> busca);
+float remocaoUnorderedMap(unordered_map<double, Record> *hash,vector<double> busca);
 
-long insercaoVetor(vector<double> *vetor,vector<double> numeros);
-long buscaVetor(vector<double> *vetor,vector<double> busca);
-long remocaoVetor(vector<double> *vetor,vector<double> busca);
+float insercaoVetor(vector<double> *vetor,vector<double> numeros);
+float buscaVetor(vector<double> *vetor,vector<double> busca);
+float remocaoVetor(vector<double> *vetor,vector<double> busca);
 
-long insercaoMap(map<double,Record> *map,vector<double> numeros, int x);
-long buscaMap(map<double, Record> *map,vector<double> busca);
-long remocaoMap(map<double, Record> *map,vector<double> busca);
+float insercaoMap(map<double,Record> *map,vector<double> numeros, int x);
+float buscaMap(map<double, Record> *map,vector<double> busca);
+float remocaoMap(map<double, Record> *map,vector<double> busca);
 
 #endif
